+++ cmd/daily.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/devOpifex/cranlogs/color"
-	"github.com/devOpifex/cranlogs/data"
+	"github.com/devOpifex/cranlogs/internal/color"
+	"github.com/devOpifex/cranlogs/internal/data"
 	"github.com/spf13/cobra"
 )
