+++ internal/color/color.go
