+++ internal/data/daily.go
